requirements.txt
streamlit
google-generativeai
langchain
pdfplumber
langchain-community
faiss-cpu
langchain-google-genai
python-dotenv